ai-server/requirements.txt
# Morrowind AI Framework - Requirements

# Core dependencies
websockets>=10.3
python-dotenv>=0.20.0
tenacity>=8.0.1

# LLM providers
openai==0.28.0
anthropic>=0.2.8
# Removed llama-cpp dependencies as they're optional

# Voice providers
elevenlabs>=0.2.18
pyttsx3>=2.90

# Utilities
aiohttp>=3.8.3
numpy>=1.22.0
pydantic>=1.9.0
